requirements.txt
Django
django-cors-headers
djangorestframework
joblib
lightgbm
numpy
pandas==1.4.2
scikit-learn==1.0.2
scipy
django-sslserver
torch
torchvision

